requirements.txt
pynput==1.8.1
requests==2.32.5
python-dotenv==1.1.1